Minor clean-ups to aid future SPARK work
SipFinalization was using an 'in out' parameter rather than 'in' but the
state was never used again in practice. SArray_Tail_To_U64_LE was using
an array of the form 'for I of S' rather than 'for I in S'Range' but the
latter is not much more verbose. In both cases, switching usage is not
hard and allows future SPARK usage.