Move Initial_State into body
SPARK GPL 2015 doesn't seem to want to accept having a 'Part_Of' aspect
on Initial_State that links it to the abstract global state in the
public part of the package. It keeps complaining this is violating RM
7.2.6 (5) but 7.2.6 (2) seems to require this aspect...

To make progress, Initial_State was moved to the body, and a new
Get_Initial_State private function has been created so child packages
can initialise their state.