Prove package SipHash
The root SipHash package now completes flow analysis and proof with no
errors and no warnings. This required re-writing the loop indexes to
make it easier to come up with loop invariants.

The Tail_To_U64_LE functions were also modified to only take non-empty
arrays and not to write the message length mod 256 to the final byte of
the padding. The main SipHash function now does this regardless of
whether the final block is a fully empty padding block or not.

SipHash.Discrete was modified in the same way, although this has not yet
been proved.