Use Compile_Time_Error for instantiation limits
Some assertions were being to make sure that the generic units were not
being instantiated in ways that would make them fail - for example if
Storage_Element was not a byte. These have been changed to the GNAT
pragma Compile_Time_Error. While not an Ada standard, it is a much
better fit for this usage.