Prove a slightly modified SipHash.General
The fact that SipHash.General uses Ada.Storage_IO is a significant
problem because Storage_IO is not really compatible with SPARK at all.
The solution found was to make a separate version of SipHash.General
called SipHash.General_SPARK (in the general-provable directory)
which takes in the buffer size and a Write procedure as generic
parameters rather than using Storage_IO.

A dummy Integer-Only non-generic version of Storage_IO was created
(it is probably not very portable across platforms) so that a gnatprove
could be run against a test instantiation in SipHash24.Integer_Hash.

Running diff across siphash-general.adb and
general-provable/siphash-general_spark.adb shows that there are
minimal differences, so a proof on the latter should give confidence
about the former.