Add gprbuild project files
There are two project files provided to build the library and the
accompanying examples.

Switch "mode" can be set to "debug" or "optimize" to control the
compilation settings The settings "analyse"/"analyze"  include the
source directories with the additional special-case sources - this is
useful when the file is being passed to gnatprove.

Switch "entropy" can be set to "urandom" or "none" to control which
system entropy source is compiled in.