Add Set_Key_From_System_Entropy with a result var
A version of Set_Key_From_System_Entropy that returns a Boolean to
indicate success rather than raising an exception has been added. This
is more in line with SPARK, where exceptions cannot be handled. It would
have been nice to make this a function, but SPARK functions cannot have
side effects.