Suppress unnecessary warnings from GNATprove
GNATprove was issuing warnings for use of the Compile_Time_Error pragma
being used to check that standard types were compatible. These checks
are not proved by GNATprove but it does not seem useful to generate a
warning about this each time.

Another warning related to the need for Integer_Storage_IO to reference
the abstract state Heap for reasons that are unclear - this was resolved
by changing the reference to an Input only.