Remove Storage_Array length overflow for SPARK Discovery GPL 2017
In SPARK Discovery GPL 2017 Storage_Array are no longer assumed to have
a length that fits into a Long_Long_Integer. This means that based on
the definition of Storage_Offset, there is no machine representation
that can actually hold the 'Length of a general Storage_Array.
Arithmetic on 'First and 'Last also produces overflow problems.

In order to prevent the subroutine preconditions from causing errors, it
is possible to use a guard expression to check that the length of the
array will not overflow.